# docs/swagger.yaml
basePath: /federation/v1
definitions:
...
  models.ComputeResourceInfo:
    properties:
      cpuArchType:
        description: CPU Instruction Set Architecture (ISA) E.g., Intel, Arm etc.
        type: string
      cpuExclusivity:
        description: Support for exclusive CPUs
        type: boolean
      diskStorage:
        description: Amount of disk storage in Gbytes for a given ISA type
        type: integer
      fpga:
        description: Number of FPGAs available for a given ISA type
        type: integer
      gpu:
        items:
          $ref: '#/definitions/models.GpuInfo'
        type: array
      hugepages:
        items:
          $ref: '#/definitions/models.HugePage'
        type: array
      memory:
        description: Amount of RAM in Mbytes
        type: integer
      numCPU:
        type: string
      vpu:
        description: Number of Intel VPUs available for a given ISA type
        type: integer
    type: object
  models.CpuArchType:
    enum:
    - ISA_X86
    - ISA_X86_64
    - ISA_ARM_64
    type: string
    x-enum-varnames:
    - X86
    - X86_64
    - ARM_64
...
  models.Flavour:
    properties:
      cpuArchType:
        $ref: '#/definitions/models.CpuArchType'
      cpuExclusivity:
        description: Support for exclusive CPUs
        type: boolean
      flavourId:
        type: string
      fpga:
        description: Number of FPGAs
        type: integer
      gpu:
        items:
          $ref: '#/definitions/models.GpuInfo'
        type: array
      hugepages:
        items:
          $ref: '#/definitions/models.HugePage'
        type: array
      memorySize:
        description: Amount of RAM in Mbytes
        type: integer
      numCPU:
        description: Number of available vCPUs
        type: integer
      storageSize:
        description: Amount of disk storage in Gbytes
        type: integer
      supportedOSTypes:
        description: A list of operating systems which this flavour configuration
          can support e.g., RHEL Linux, Ubuntu 18.04 LTS, MS Windows 2012 R2.
        items:
          $ref: '#/definitions/models.OsType'
        type: array
      vpu:
        description: Number of Intel VPUs available
        type: integer
    type: object
  models.GpuInfo:
    properties:
      gpuMemory:
        description: GPU memory in Mbytes
        type: integer
      gpuModeName:
        description: Model name corresponding to vendorType may include info e.g.
          for NVIDIA, model name could be “Tesla M60”, “Tesla V100” etc.
        type: string
      gpuVendorType:
        description: GPU vendor name e.g. NVIDIA, AMD etc.
        type: string
      numGPU:
        description: Number of GPUs
        type: integer
    type: object
  models.HugePage:
    properties:
      number:
        description: Total number of huge pages
        type: integer
      pageSize:
        description: Size of hugepage
        type: string
    type: object
  models.InvalidParam:
    properties:
      param:
        type: string
      reason:
        type: string
    type: object
...
  models.OsType:
    properties:
      architecture:
        type: string
      distribution:
        type: string
      license:
        type: string
      version:
        type: string
    type: object
  models.ProblemDetails:
    properties:
      cause:
        description: Fixed string indicating cause of the issue
        type: string
      detail:
        description: Specific detail of the issue
        type: string
      invalidParams:
        items:
          $ref: '#/definitions/models.InvalidParam'
        type: array
      title:
        description: Summary of the problem
        type: string
    type: object
...
  models.ZoneDetails:
    properties:
      geographyDetails:
        description: Details about cities or state covered by the edge. Details about
          the type of locality for eg rural, urban, industrial etc. This information
          is defined in human readable form.
        type: string
      geolocation:
        type: string
      vimId:
        description: vim id
        type: string
      zoneId:
        description: cluster id
        type: string
    type: object
  models.ZoneRegisteredData:
    properties:
      computeResourceQuotaLimits:
        description: Max quota on resources partner OP allows over reserved resources.
        items:
          $ref: '#/definitions/models.ComputeResourceInfo'
        type: array
      federationContextId:
        type: string
      flavoursSupported:
        items:
          $ref: '#/definitions/models.Flavour'
        type: array
      networkResources: {}
      reservedComputeResources:
        description: Resources exclusively reserved for the originator OP.
        items:
          $ref: '#/definitions/models.ComputeResourceInfo'
        type: array
      zoneId:
        type: string
      zoneServiceLevelObjsInfo:
        description: It is a measure of the actual amount of data that is being sent
          over a network per unit of time and indicates máximum supported value for
          a zone
    type: object
  models.ZoneRegistrationRequestData:
    properties:
      acceptedAvailabilityZones:
        items:
          type: string
        type: array
      availZoneNotifLink:
        type: string
    type: object
  models.ZoneRegistrationResponseData:
    properties:
      acceptedZoneResourceInfo:
        items:
          $ref: '#/definitions/models.ZoneRegisteredData'
        type: array
    type: object
host: localhost:8000
info:
  contact: {}
  description: This is the API documentation for the MEC Federator.
  title: MEC Federator API
  version: "1.0"
paths:
...
  /ewbi/{federationContextId}/zones:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all availability zones offered by the partner
        operator
      parameters:
      - description: Federation Context ID
        in: path
        name: federationContextId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ZoneDetails'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ProblemDetails'
      summary: Get All Local Zones
      tags:
      - EWBI - ZonesInfoSync
    post:
      consumes:
      - application/json
      description: Registers the origin operator's intent to use one or more availability
        zones from the partner operator
      parameters:
      - description: Federation Context ID
        in: path
        name: federationContextId
        required: true
        type: string
      - description: Zone Registration Request Data
        in: body
        name: zoneRegistrationRequestData
        required: true
        schema:
          $ref: '#/definitions/models.ZoneRegistrationRequestData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ZoneRegistrationResponseData'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/models.ProblemDetails'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ProblemDetails'
      summary: Subscribe to Zone
      tags:
      - EWBI - ZonesInfoSync
  /ewbi/{federationContextId}/zones/{zoneId}:
    delete:
      consumes:
      - application/json
      description: Unregisters the origin operator's subscription to a specific availability
        zone from the partner operator
      parameters:
      - description: Federation Context ID
        in: path
        name: federationContextId
        required: true
        type: string
      - description: Zone identifier
        in: path
        name: zoneId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Zone unregistered successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ProblemDetails'
      summary: Unsubscribe from Zone
      tags:
      - EWBI - ZonesInfoSync
    get:
      consumes:
      - application/json
      description: Retrieves detailed information about a specific availability zone
        from the partner operator
      parameters:
      - description: Federation Context ID
        in: path
        name: federationContextId
        required: true
        type: string
      - description: Zone identifier
        in: path
        name: zoneId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ZoneDetails'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ProblemDetails'
      summary: Get Zone Details
      tags:
      - EWBI - ZonesInfoSync
...
swagger: "2.0"